Add files via upload
The products in the category were divided into five grades by price, and the products purchased by the customer were replaced with grades. If the customer bought a lot, the product was graded and averaged. On average, we gave larger scores to customers who bought a large number of high-grade products and performed a linear regression models.